requirements.txt
deepgram-sdk==3.8.0
python-dotenv==1.0.0
pyaudio==0.2.14
sounddevice==0.4.6
numpy==1.26.3

